+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a `Batch#submit` method to submit directly with a `qsub` call.
+  [#29](https://github.com/OSC/pbs-ruby/issues/29)
+
 ### Changed
 - Changed the `CHANGELOG.md` formatting.
 - Updated date in `LICENSE.txt`.